refactor(admin): centralize action auth wrapper
Move admin session validation into app/lib/with-auth.ts so root actions and scoped action modules share one auth wrapper without lib/with-auth importing back through the root action barrel.

Verified with pnpm --filter @anipotts/admin typecheck, pnpm --filter @anipotts/admin lint, pnpm --filter @anipotts/admin build, pnpm exec prettier --check apps/admin/src/app/actions.ts apps/admin/src/app/lib/with-auth.ts, git diff --check -- apps/admin/src/app/actions.ts apps/admin/src/app/lib/with-auth.ts, and pnpm dlx knip --no-exit-code.
